⚜️ Rename ICompositionRequestsHandler impls to *Composer
NServiceBus's IHandleMessages<T> and ServiceComposer's
ICompositionRequestsHandler both produced classes ending in Handler,
so a reader had to peek inside to know which world a class belonged
to. Renaming the ServiceComposer side to *Composer disambiguates the
two and matches the library that owns them. ICompositionEventsSubscriber
impls stay as *Subscriber; NServiceBus IHandleMessages<T> impls keep
Handler.
