+++ src/Catalog.ServiceComposition/Checkout/ShoppingCartComposer.cs
@@ -1,6 +1,4 @@
-using System.Dynamic;
 using Catalog.Data;
-using Catalog.Data.Models;
 using Catalog.ServiceComposition.Events;
 using Catalog.ServiceComposition.Helpers;
 using Microsoft.AspNetCore.Http;
@@ -11,9 +9,16 @@ using ServiceComposer.AspNetCore;
 
 namespace Catalog.ServiceComposition.Checkout;
 
-public class SummaryHandler(CatalogDbContext dbContext, CacheHelper cacheHelper) : ICompositionRequestsHandler
+// Powers the shipping screen, which renders the full `CartItemList`
+// (product image + name) on top of the order summary. Reads from the
+// SQLite Orders row first; if the SubmitOrderItems message hasn't been
+// processed yet, falls back to the in-flight cart in the distributed
+// cache so the page renders cleanly while the read model catches up.
+//
+// Address and payment use the leaner `CartSummaryComposer` instead.
+public class ShoppingCartComposer(CatalogDbContext dbContext, CacheHelper cacheHelper) : ICompositionRequestsHandler
 {
-    [HttpGet("/buy/summary/{orderId}")]
+    [HttpGet("/buy/shipping/{orderId}")]
     public async Task Handle(HttpRequest request)
     {
         var orderIdString = (string)request.HttpContext.GetRouteData().Values["orderId"]!;
@@ -29,33 +34,17 @@ public class SummaryHandler(CatalogDbContext dbContext, CacheHelper cacheHelper)
             order = await cacheHelper.GetOrder(orderId);
         }
 
-        var productsModel = MapToDictionary(order.Products);
+        var products = await dbContext.Products.ToListAsync(ct);
+        var orderedProducts = ShoppingCart.Mapper.MapToDictionary(order, products);
 
         var context = request.GetCompositionContext();
-        await context.RaiseEvent(new SummaryLoaded
+        await context.RaiseEvent(new CartLoaded
         {
-            OrderId = orderId,
-            Products = productsModel
+            OrderedProducts = orderedProducts
         });
 
         var vm = request.GetComposedResponseModel();
         vm.OrderId = orderId;
-        vm.Products = productsModel;
-    }
-
-    IDictionary<Guid, dynamic> MapToDictionary(List<OrderItem> products)
-    {
-        var productsViewModel = new Dictionary<Guid, dynamic>();
-
-        foreach (var product in products)
-        {
-            dynamic vm = new ExpandoObject();
-            vm.ProductId = product.ProductId;
-            vm.Quantity = product.Quantity;
-
-            productsViewModel[product.ProductId] = vm;
-        }
-
-        return productsViewModel;
+        vm.CartItems = orderedProducts.Values.ToList();
     }
 }
