+++ src/Finance.ServiceComposition/Checkout/DeliveryOptionSubmitComposer.cs
@@ -1,22 +1,17 @@
-﻿using Microsoft.AspNetCore.Http;
+﻿using Finance.Endpoint.Messages.Commands;
+using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using ServiceComposer.AspNetCore;
-using Shipping.Endpoint.Messages.Commands;
-using Shipping.ServiceComposition.Helpers;
 
-namespace Shipping.ServiceComposition.Checkout;
+namespace Finance.ServiceComposition.Checkout;
 
-public class DeliveryOptionSubmitHandler(IMessageSession messageSession, CacheHelper cacheHelper) : ICompositionRequestsHandler
+public class DeliveryOptionSubmitComposer(IMessageSession messageSession) : ICompositionRequestsHandler
 {
     [HttpPost("/buy/shipping/{orderId}")]
     public async Task Handle(HttpRequest request)
     {
         var submitted = await request.Bind<SelectedDeliveryOption>();
 
-        var order = await cacheHelper.GetOrder(submitted.OrderId);
-        order.DeliveryOptionId = submitted.Body.DeliveryOptionId;
-        await cacheHelper.StoreOrder(order);
-        
         var message = new SubmitDeliveryOption()
         {
             OrderId = submitted.OrderId,
