+++ src/Finance.ServiceComposition/Checkout/OrderSubmitComposer.cs
@@ -11,7 +11,7 @@ using OrderItem = Finance.Endpoint.Messages.Commands.OrderItem;
 
 namespace Finance.ServiceComposition.Checkout;
 
-public class OrderSubmitHandler(IMessageSession messageSession, CacheHelper cacheHelper) : ICompositionRequestsHandler
+public class OrderSubmitComposer(IMessageSession messageSession, CacheHelper cacheHelper) : ICompositionRequestsHandler
 {
     [HttpPost("/cart/{orderId}")]
     public async Task Handle(HttpRequest request)
