+++ src/Shipping.ServiceComposition/Checkout/AddressSubmitComposer.cs
@@ -6,7 +6,7 @@ using Shipping.Endpoint.Messages.Commands;
 
 namespace Shipping.ServiceComposition.Checkout;
 
-public class AddressSubmitHandler(IMessageSession messageSession) : ICompositionRequestsHandler
+public class AddressSubmitComposer(IMessageSession messageSession) : ICompositionRequestsHandler
 {
     [HttpPost("/buy/address/{orderId}")]
     public async Task Handle(HttpRequest request)
